# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kasir API",
        "description": "Dokumentasi API untuk aplikasi kasir.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base URL"
        }
    ],
    "paths": {
...
        "/transaksi": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Daftar transaksi toko",
                "operationId": "ef4e2cf57d8d7917aed6e29f970c156e",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Cari kode transaksi",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar transaksi berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Buat transaksi baru",
                "operationId": "7bfde9e8fc8cd98e509cf10ec6d4bce4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items",
                                    "metode_pembayaran",
                                    "nominal_bayar"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "produk_id": {
                                                    "type": "string",
                                                    "example": "01kty..."
                                                },
                                                "jumlah": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "metode_pembayaran": {
                                        "type": "string",
                                        "example": "cash"
                                    },
                                    "nominal_bayar": {
                                        "type": "integer",
                                        "example": 50000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaksi berhasil dibuat"
                    },
                    "400": {
                        "description": "Permintaan tidak valid"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/transaksi/{transaksi}": {
            "get": {
                "tags": [
                    "Transaksi"
                ],
                "summary": "Detail transaksi",
                "operationId": "124fb87792734f8101f10d733c4c2eaf",
                "parameters": [
                    {
                        "name": "transaksi",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail transaksi berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan bearer token hasil login.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}